upadted changes in project

+++ HRMS/src/app/modules/master/designation/designation-add/designation-list/designation-list.component.ts
@@ -9,6 +9,7 @@ import {
   faTrash,
 } from '@fortawesome/free-solid-svg-icons';
 import { NgxUiLoaderService } from 'ngx-ui-loader';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-designation-list',
@@ -30,7 +31,7 @@ export class DesignationListComponent implements OnInit {
   constructor(
     public _designationService: DesignationService,
     private _notificationService: NotificationService,
-    private ngxService: NgxUiLoaderService
+    private loaderService: NgxUiLoaderService
   ) {}
 
   ngOnInit(): void {}
@@ -43,20 +44,36 @@ export class DesignationListComponent implements OnInit {
 
   //method to delete
   onDelete(designation: IDesignation) {
-    this.ngxService.start();
+    Swal.fire({
+      title: 'Are you sure want to delete?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it',
+    }).then((result: any) => {
+      if (result.value) {
+        // Swal.fire('Deleted!', 'Designation has been deleted.', 'success');
+        this.loaderService.start();
 
-    //call method from service to delete designation
-    this._designationService.deleteDesignation(designation);
+        //call method from service to delete designation
+        this._designationService.deleteDesignation(designation);
 
-    //toaster notification on success from notificationService
-    this._notificationService.onSuccess(
-      'Designation deleted successfully',
-      'Notification'
-    );
+        //toaster notification on success from notificationService
+        this._notificationService.onSuccess(
+          'Designation deleted successfully',
+          'Notification'
+        );
 
-    console.log(this._designationService.designations);
+        if (this._designationService.designations.length === 0)
+          this.isNoData = true;
 
-    if (this._designationService.designations.length === 0)
-      this.isNoData = true;
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        // Swal.fire('Cancelled', 'Designation is safe :)', 'error');
+        this._notificationService.onError(
+          'Designation is safe :)',
+          'Notification'
+        );
+      }
+    });
   }
 }
