+++ HRMS/src/app/modules/master/external-designation/external-designation-add/external-designation-list/external-designation-list.component.ts
@@ -9,6 +9,7 @@ import {
   faTrash,
 } from '@fortawesome/free-solid-svg-icons';
 import { NgxUiLoaderService } from 'ngx-ui-loader';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-external-designation-list',
@@ -16,7 +17,6 @@ import { NgxUiLoaderService } from 'ngx-ui-loader';
   styleUrls: ['./external-designation-list.component.css'],
 })
 export class ExternalDesignationListComponent implements OnInit {
-
   //properties for icons
   faToggleOn = faToggleOn;
   faToggleOff = faToggleOff;
@@ -31,35 +31,54 @@ export class ExternalDesignationListComponent implements OnInit {
   constructor(
     public _externalDesignationService: ExternalDesignationService,
     private _notificationService: NotificationService,
-    private ngxService: NgxUiLoaderService
+    private loaderService: NgxUiLoaderService
   ) {}
 
   ngOnInit(): void {}
 
   // method to edit
   onEdit(externalDesignation: IDesignation) {
-
     // emit designation when click on edit
     this.sendExternalDesignation.emit(externalDesignation);
   }
 
   // method to delete
   onDelete(externalDesignation: IDesignation) {
+    Swal.fire({
+      title: 'Are you sure want to delete?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it',
+    }).then((result) => {
+      if (result.value) {
+        // Swal.fire(
+        //   'Deleted!',
+        //   'External designation has been deleted.',
+        //   'success'
+        // );
+        this.loaderService.start();
 
-    this.ngxService.start();
-    
-    // call method for delete designation
-    this._externalDesignationService.deleteExternalDesignation(
-      externalDesignation
-    );
+        // call method for delete designation
+        this._externalDesignationService.deleteExternalDesignation(
+          externalDesignation
+        );
 
-    // toaster notification on success
-    this._notificationService.onSuccess(
-      'External designation deleted successfully ',
-      'notification'
-    );
+        // toaster notification on success
+        this._notificationService.onSuccess(
+          'External designation deleted successfully ',
+          'notification'
+        );
 
-    if(this._externalDesignationService.externalDesignations.length === 0)
-    this.isNoData = true;
+        if (this._externalDesignationService.externalDesignations.length === 0)
+          this.isNoData = true;
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        // Swal.fire('Cancelled', 'External designation is safe :)', 'error');
+        this._notificationService.onError(
+          'External designation is safe :)',
+          'notification'
+        );
+      }
+    });
   }
 }
